+++ cmsplugin_blocks/utils.py
@@ -4,7 +4,7 @@ import zipfile
 
 from io import BytesIO
 
-# This is probably for Python 2/3 support although we dont support Py2
+# For Python 2/3 support
 try:
     import Image as PILImage
 except ImportError:
@@ -29,7 +29,7 @@ def is_valid_image_filename(filename):
     """
     ext = filename.split('.')[-1]
 
-    if ext.lower() in settings.BLOCKS_MASSUPLOAD_IMAGE_TYPES:
+    if ext.lower() in settings.BLOCKS_ALLOWED_IMAGE_EXTENSIONS:
         return True
 
     return False
@@ -40,13 +40,15 @@ def validate_file_size(data):
     Validate file size does not exceed limit from
     ``settings.BLOCKS_MASSUPLOAD_FILESIZE_LIMIT``.
 
-    This returns nothing.
-
     Raises:
         ValidationError: If file size is over limit.
 
     Arguments:
         data (file object):
+
+    Returns:
+        boolean: Always True, obvisously excepted if an exception is raised
+        when file is over limit.
     """
     msg = _('Please keep filesize under {}. Current filesize {}')
 
@@ -55,6 +57,8 @@ def validate_file_size(data):
             filesizeformat(settings.BLOCKS_MASSUPLOAD_FILESIZE_LIMIT),
             filesizeformat(data._size)
         ))
+
+    return True
 
 
 def validate_zip(data, obj=None):
@@ -178,3 +182,51 @@ def store_images_from_zip(instance, zip_fileobject, item_model,
         zip_fileobject.close()
 
     return stored_items
+
+
+class SmartFormatMixin(object):
+    """
+    A mixin to inherit from a model so it will have some common helper
+    methods to manage image formats.
+    """
+    def media_format(self, mediafile):
+        """
+        Common method to perform a naive check about image format using file
+        extension.
+
+        This has been done for common image formats, so it will return either
+        'JPEG', 'PNG', 'SVG' or None if it does not match any of these formats.
+
+        Obviously, since it use the file extension, found format is not to be
+        100% trusted. For sanity, media saving should validate it correclty
+        and possibly enforce the right file extension according to file format
+        found from file metas (like with the PIL method to get it).
+
+        At least the FileField should validate than uploaded file is a valid
+        image.
+
+        Why using this naive technique ? To be able to use it in templates
+        without cache and so avoid to open image file each time template
+        is rendered.
+
+        Arguments:
+            mediafile (object): Either a FileField or ImageField or any other
+                object which implement a ``name`` attribute which return the
+                filename.
+
+        Return:
+            string: Format name if filename extension match to any available
+            format extension, else ``None``.
+        """
+        if mediafile:
+            ext = mediafile.name.split(".")[-1].lower()
+            if ext in ["jpg", "jpeg"]:
+                return "JPEG"
+            elif ext in ["png"]:
+                return "PNG"
+            elif ext in ["gif"]:
+                return "GIF"
+            elif ext in ["svg"]:
+                return "SVG"
+
+        return None