+++ HISTORY.rst
@@ -9,6 +9,7 @@ Version 0.6.0 - Unreleased
 * Modified every plugin templates to be cleaner and flawless;
 * Added ``title`` attribute to SlideItem model;
 * Use **sorl-thumbnail** on images in every plugin templates;
+* Use FileField instead of ImageField for image fields in every plugin;
 
 Version 0.5.2 - 2019/05/18
 --------------------------